[11/18] 招待機能　エラーあり

+++ app/controllers/public/memberships_controller.rb
@@ -1,5 +1,4 @@
 class Public::MembershipsController < ApplicationController
-  before_action :authenticate_user!
 
   def create
     @family = Family.find(params[:family_id])
@@ -9,9 +8,26 @@ class Public::MembershipsController < ApplicationController
   end
 
   def destroy
+    @family = Family.find(params[:family_id])
     membership = current_user.memberships.find_by(family_id: params[:family_id])
-    membership.destroy
-    redirect_to request.referer, alert: "グループへの参加申請を取消しました"
+    if membership
+      membership.destroy
+      redirect_to request.referer, alert: "グループへの参加申請を取消しました"
+    else
+      # メンバーシップが見つからない場合の処理
+      redirect_to family_path(@family), alert: "メンバーシップが見つかりませんでした"
+    end
+
+    # @membership = Membership.find(params[:id])
+    # @membership.destroy!
+    # @family = Family.find(params[:family_id])
+    # redirect_to family_url(@family), notice: "加入申請を取り消しました"
+  end
+
+  private
+
+  def membership_params
+    params.permit(:family_id)
   end
 
 end
